change back to size_t

+++ utils/include/utils/ConfigurationUtils.h
@@ -16,7 +16,7 @@
  */
 #pragma once
 
-#include <cstdint>
+#include <cstddef>
 
 namespace org::apache::nifi::minifi {
 
@@ -24,8 +24,8 @@ class Configure;
 
 namespace utils::configuration {
 
-inline constexpr uint64_t DEFAULT_BUFFER_SIZE = 4096;
-uint64_t getBufferSize(const Configure& configuration);
+inline constexpr size_t DEFAULT_BUFFER_SIZE = 4096;
+size_t getBufferSize(const Configure& configuration);
 
 }  // namespace utils::configuration
 }  // namespace org::apache::nifi::minifi
